# [#3932] Use :npm-deps false to speed up Figwheel startup