fixes #3932

### Summary:
Speed up clojurescript compilation 2 times.

```
Successfully compiled build :android to “target/android/app.js" in 144.678 seconds.
Successfully compiled build :android to "target/android/app.js" in 64.963 seconds.

Successfully compiled build :ios to "target/ios/app.js" in 133.19 seconds.
Successfully compiled build :ios to "target/ios/app.js" in 73.528 seconds.
```

status: ready