@dmitryn 2018-05-01
Compiled the project and ran it in iOS simulator - works fine for me.
Not sure manual testing is required here, so we could wait for e2e tests to pass and then merge.